requirements.txt
# Mobius Platform - Consolidated Python Dependencies
# Usage: python3.11 -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt
#
# Last updated: 2026-02-06

# =============================================================================
# Web Frameworks
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
flask>=3.0.0
python-multipart>=0.0.6

# =============================================================================
# Database (PostgreSQL)
# =============================================================================
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.0      # Sync driver (Mobius-user, mobius-os)
psycopg2-binary>=2.9.0      # Sync driver (mobius-chat, mobius-dbt)
asyncpg>=0.29.0             # Async driver (mobius-rag)
alembic>=1.13.0             # Migrations
greenlet>=3.0.0             # SQLAlchemy async support

# =============================================================================
# Google Cloud Platform
# =============================================================================
google-cloud-aiplatform>=1.40.0   # Vertex AI (LLM, embeddings)
google-cloud-storage>=2.14.0      # GCS buckets
google-api-python-client>=2.0.0   # Google Drive API (mobius-rag)
google-auth-oauthlib>=1.0.0       # OAuth for Drive (mobius-rag)
google-cloud-bigquery>=3.0.0      # BigQuery (mobius-dbt)
google-cloud-firestore>=2.14.0    # Firestore (mobius-os)

# =============================================================================
# Redis (Queuing)
# =============================================================================
redis>=5.0.0

# =============================================================================
# Authentication & Security
# =============================================================================
bcrypt>=4.1.0
PyJWT>=2.8.0

# =============================================================================
# HTTP & Networking
# =============================================================================
httpx>=0.26.0
aiohttp>=3.9.0

# =============================================================================
# Data Processing
# =============================================================================
pydantic>=2.0.0
pyyaml>=6.0.0
json-repair>=0.55.0         # Fix malformed LLM JSON

# =============================================================================
# Mobius Skills - Google Search (DuckDuckGo fallback for doc assembly)
# =============================================================================
ddgs>=9.0.0

# =============================================================================
# MCP (Model Context Protocol) - mobius-skills-mcp server
# =============================================================================
mcp[cli]>=1.0.0

# =============================================================================
# Document Processing (RAG)
# =============================================================================
pymupdf>=1.23.0             # PDF parsing
beautifulsoup4>=4.12.0      # HTML parsing

# =============================================================================
# DBT (BigQuery Transformations)
# =============================================================================
dbt-bigquery>=1.7.0,<3.0.0

# =============================================================================
# Environment & Config
# =============================================================================
python-dotenv>=1.0.0
pytz>=2024.1

# =============================================================================
# Testing
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0

# =============================================================================
# Mobius-user (install as editable after pip install)
# Run: pip install -e ./Mobius-user
# =============================================================================
